rename modDefinitionsHooks -> modDefinitionHooks
