+++ src/modDefinitions/modDefinitionHooks.ts
