+++ src/pageEditor/sidebar/DynamicEntry.test.tsx
@@ -23,7 +23,7 @@ import { authActions } from "@/auth/authSlice";
 import { formStateFactory } from "@/testUtils/factories/pageEditorFactories";
 import { authStateFactory } from "@/testUtils/factories/authFactories";
 
-jest.mock("@/modDefinitions/modDefinitionsHooks", () => ({
+jest.mock("@/modDefinitions/modDefinitionHooks", () => ({
   useAllModDefinitions: jest
     .fn()
     .mockReturnValue({ data: [], isLoading: false }),
