+++ src/pageEditor/sidebar/SidebarExpanded.tsx
@@ -57,7 +57,7 @@ import AddExtensionPointButton from "./AddExtensionPointButton";
 import ExtensionEntry from "./ExtensionEntry";
 import { actions } from "@/pageEditor/slices/editorSlice";
 import { measureDurationFromAppStart } from "@/utils/performance";
-import { useAllModDefinitions } from "@/modDefinitions/modDefinitionsHooks";
+import { useAllModDefinitions } from "@/modDefinitions/modDefinitionHooks";
 import { useDebounce } from "use-debounce";
 
 const SidebarExpanded: React.FunctionComponent<{
